requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.38.0
-e agents/prompt_builder/gpt-cost-estimator
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lorem-text==2.1
matplotlib==3.7.0
networkx==3.0
numpy==1.24.2
omegaconf==2.3.0
openai==1.35.14
packaging==23.0
Pillow==9.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.1.3
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
sniffio==1.3.1
tifffile==2023.2.3
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.14.0
